+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.5.0-rc.7",
+  sdkVersion: "4.5.0-rc.8",
   genVersion: "2.841.0",
-  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.7 2.841.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.8 2.841.0 2.0 @orq-ai/node",
 } as const;